Refactor unistore actions to redux-toolkit actions and reducer
